refactor(app): extract BattleRoutes from App.tsx (#316 phase 8b)
The `playing` and `battlesummary` screens, including the phase dispatch
inside `playing` — celebration/VictoryPanel, fingerSmash, waveReward,
gameOver/GameOver, and the live Battlefield with BossShockwave and the nine
rare-event overlays.

BattleContextValue grows past the reducer triple to carry what the battle
render actually reads: showBossSplash, actTheme, isCampaign,
quickPlayRewardClaimed, the rare-event pair, the battle controls, the
post-battle handlers, and the six mode refs (isCampaignRef,
worldBattleNodeIdRef, isDailyChallengeRef, isWeeklyChallengeRef,
gameStateRef, summaryDoneRef).

Those stay refs rather than becoming state: the battle flow mutates them
synchronously between screens, and BattleRoutes re-renders exactly when App
did, so each read happens at the same point it used to.

One removal beyond a pure move: the `fingerSmash` branch bound
`const fp = gameState.phase as {...}` and never used it — dead on main too
(tsc --noUnusedLocals confirms), and it would trip the same check in the new
file. Dropped rather than carried over.

App.tsx: 2,995 -> 2,894 lines. Build green, chunk boundaries unchanged,
2107 tests passing.
