+++ web/src/app/BattleContext.tsx
@@ -1,6 +1,7 @@
-import { createContext, useContext, type Dispatch, type ReactNode } from 'react'
-import type { GameState } from '../game/types'
+import { createContext, useContext, type Dispatch, type MutableRefObject, type ReactNode, type SetStateAction } from 'react'
+import type { GameState, StanceRules } from '../game/types'
 import type { BattleAction, BattleState } from '../game/battleReducer'
+import type { RareEventKind } from '../components/rare-events/types'
 
 /**
  * Battle-only state, kept out of AppContext on purpose: `gameState` is replaced
@@ -13,7 +14,49 @@ export interface BattleContextValue {
   battle:    BattleState
   gameState: GameState | null
   dispatch:  Dispatch<BattleAction>
+
+  showBossSplash:        boolean
+  actTheme:              string | undefined
+  isCampaign:            boolean
+  quickPlayRewardClaimed: boolean
+  activeRareEvent:       RareEventKind | null
+  handleRareEventDone:   () => void
+
+  /**
+   * Mode flags read during render. They are refs rather than state because the
+   * battle flow mutates them synchronously between screens; a consumer reads
+   * them at the same points App used to.
+   */
+  isCampaignRef:         MutableRefObject<boolean>
+  worldBattleNodeIdRef:  MutableRefObject<string | null>
+  isDailyChallengeRef:   MutableRefObject<boolean>
+  isWeeklyChallengeRef:  MutableRefObject<boolean>
+  gameStateRef:          MutableRefObject<GameState | null>
+  summaryDoneRef:        MutableRefObject<() => void>
+
+  // ── Battle controls ───────────────────────────────────────────────────────
+  handlePlayCard:    (cardId: string) => void
+  handlePlayAoeCard: (cardId: string, cx: number, cy: number) => void
+  handleGiveUp:      () => void
+  setIsUserPaused:   Dispatch<SetStateAction<boolean>>
+  handleSetStance:   (s: NonNullable<GameState['playerStance']>) => void
+  handleCycleSpeed:  () => void
+  handleWaveRewardPick: (cardName: string) => void
+  handleWaveRewardSkip: () => void
+
+  // ── Post-battle ───────────────────────────────────────────────────────────
+  handleOpenPack:            () => void
+  handleCampaignWin:         () => void
+  handleCampaignRetry:       () => void
+  handleDailyChallengeRetry: () => void
+  handlePlayAgain:           () => void
+  handleWorldBattleRetry:    () => void
+  handleMainMenu:            () => void
+  handleAbandonRun:          () => void
 }
+
+/** Re-exported for consumers that narrow stance options per node type. */
+export type { StanceRules }
 
 const BattleContext = createContext<BattleContextValue | null>(null)
